Revert "cleanup in prep to release to mvn central"
This reverts commit c7e2ab67849c342d5f8e98d771477526a0368a13.

+++ src/main/java/colt/nicity/core/lang/MinMaxInt.java
@@ -141,7 +141,7 @@ public class MinMaxInt {
      * @return
      */
     public double mean() {
-        return sum / count;
+        return sum / (int) count;
     }
     /**
      *
